requirements.txt
streamlit
pdfplumber
python-docx
openai
pandas
openpyxl
htmldocx 
